Moved 1st_ode_linear_system to ode.py, so we now to solve a linear system of equations we can just simply call dsolve.
Updated the doctests to reflect this.
